pyproject.toml
[project]
name = "violence-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "openai-clip>=1.0.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyaml>=25.7.0",
    "requests>=2.32.5",
    "streamlit>=0.89.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
